Produce a stream of futures as a result of message
Provide a stream of futures instead of an option of them as a result of
one message. This is a preparation for the batch handling, where we'll
produce multiple futures for a single message.

Also, some cleanups in the code are included, to make all the nested
boxes of results of options manageable.